Extract definition, naming, and copular relation frames
